Updated ids to just use integers, a little easier to work with sometimes and we don't need the extra space for this. Created unit tests for the methods I created to be a help to the others.

+++ src/main/java/cs309/dto/EventDTO.java
@@ -10,7 +10,7 @@ import java.text.SimpleDateFormat;
  */
 public class EventDTO {
 
-    private Long eventId;
+    private Integer eventId;
     private String name;
     private String startTime;
     private String endTime;
@@ -51,11 +51,11 @@ public class EventDTO {
         }
     }
 
-    public Long getEventId() {
+    public Integer getEventId() {
         return eventId;
     }
 
-    public void setEventId(Long eventId) {
+    public void setEventId(Integer eventId) {
         this.eventId = eventId;
     }
