+++ src/main/java/cs309/repo/InviteRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by jeffrey on 1/31/16.
  */
-public interface InviteRepository extends JpaRepository<Invite,Long> {
+public interface InviteRepository extends JpaRepository<Invite,Integer> {
 }
